+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+* Generate upload token based on bucket and key.
+
 ## [v0.3.0](https://github.com/linjunpop/blast_off/tree/v0.3.0)
 
 * Set correct file mime-types.